fix(user): css mobile, eliminaUser
- modificato css sidebar da mobile, con relative modifiche
- modificato .admin-stats-list nel css, allargando la gap a 0.75rem per l'accessibilità
- modificata deleteRecensione: aggiunta la query censura=1 come commento, così da poterla sostituire facilmente se decidiamo di mettere quella
- modificato showEliminaUser.html e user_function.php per impedire l'eliminazione dell'account in caso ci siano prestiti attivi, e introdotto messaggio di errore in tal caso

+++ includes/functions/user_functions.php
@@ -117,9 +117,19 @@ function changeFotoProfilo($conn, $userId, $fotoPath) {
 }
 
 function deleteUtente($conn, $userId) {
+    $stmt = $conn->prepare('SELECT COUNT(*) FROM prestito WHERE utente_id = ? AND stato IN ("attivo", "in_ritardo")');
+    $stmt->bind_param('i', $userId);
+    $stmt->execute();
+    $stmt->bind_result($count);
+    $stmt->fetch();
+    $stmt->close();
+
+    if ($count > 0) {
+        return 'Impossibile eliminare l\'account se ci sono prestiti attivi';
+    }
+
     $stmt = $conn->prepare('DELETE FROM utente WHERE id = ?');
     $stmt->bind_param('i', $userId);
-
     return $stmt->execute();
 }
 ?>
